requirements.txt
fastapi[all]==0.109.2
pydantic==2.6.1
pymongo[srv]==4.6.1
python-dotenv==1.0.1
typing==3.7.4.3
uvicorn==0.27.1